% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildTestingKernels.R
\name{BuildTestingKernels}
\alias{BuildTestingKernels}
\title{Function builds Gaussian kernels for testing.}
\usage{
BuildTestingKernels(location, bandwidth)
}
\arguments{
\item{location}{A n by d matrix of cell/spot location coordinates.}

\item{bandwidth}{A numeric value of bandwidth}
}
\value{
Gaussian kernel
}
\description{
Function builds Gaussian kernels for testing.
}
