% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nngp_approx.R
\name{nngp_approx}
\alias{nngp_approx}
\title{nngp_approx}
\arguments{
\item{kernelMatrix}{the kernel matrix K from SpatialPCA}

\item{y}{gene expression matrix}

\item{m}{number of nearest number}

\item{location}{A n by d matrix of cell/spot location coordinates.}
}
\value{
sigma.inverse, sigma.inverse.det, sigma.det, F.vector
}
\description{
nngp_approx
}
