% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNGPMethod.R
\name{nngpMethodEquations_InverseKernel}
\alias{nngpMethodEquations_InverseKernel}
\title{NNGP equations}
\usage{
nngpMethodEquations_InverseKernel(
  inverseMatrix,
  F.vector,
  X,
  tau = 1,
  equation
)
}
\arguments{
\item{X}{scaled location matrix}

\item{tau}{parameter default is 1}

\item{equation}{test which equation}

\item{kernelMatrix}{the kernel matrix K}

\item{F}{vector from Nearest Neighbor Gaussian Process}
}
\value{
equation output
}
\description{
NNGP equations
}
