% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCAMethod.R
\name{SpatialPCAEquations_InverseKernel}
\alias{SpatialPCAEquations_InverseKernel}
\title{SpatialPCAMethodEquations}
\usage{
SpatialPCAEquations_InverseKernel(kernelMatrix, X, tau = 1, equation)
}
\arguments{
\item{kernelMatrix}{the kernel matrix K}

\item{X}{scaled location matrix}

\item{tau}{parameter default is 1}

\item{equation}{test which equation}
}
\value{
equation output
}
\description{
SpatialPCAMethodEquations
}
